src/deployment/api/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
torch
python-multipart